fix(notifications): properly display notification updates
