+++ CacheMoney/frontend/src/components/Transaction/Deposit.js
@@ -8,7 +8,6 @@ import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 function Deposit(props) {
-	// post deposit transaction
 	const postDeposit = (transaction) => {
 		axios
 			.post(`${config.url}accounts/deposit`, transaction, {
@@ -17,8 +16,6 @@ function Deposit(props) {
 					userId: store.getState().userReducer.userId,
 				},
 			})
-
-
 			.then(
 				result=>{
 					result.status===200?
@@ -33,7 +30,6 @@ function Deposit(props) {
 					}):toast.error('error')
 			}
 				)
-
 			.catch((error) => {
 				console.error(`Error: ${error}`)
 				toast.error('Deposit failed', {
